# Tugas/Teknologi Integrasi Sistem/UAP_TIS_TIC/backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WowoClean API - Sistem Manajemen Kontainer Limbah B3",
        "description": "Enterprise REST API untuk manajemen kontainer limbah B3 dengan autentikasi JWT, otorisasi berbasis Role, dan API Gateway pattern.",
        "contact": {
            "name": "WowoClean API Support",
            "email": "admin@wowoclean.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/gateway/containers/{containerId}/tracking-logs": {
            "get": {
                "tags": [
                    "Gateway - Tracking Logs"
                ],
                "summary": "Ambil log perjalanan kontainer",
                "description": "Mengambil semua log perjalanan untuk kontainer tertentu.",
                "operationId": "getTrackingLogs",
                "parameters": [
                    {
                        "name": "containerId",
                        "in": "path",
                        "description": "ID kontainer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Log perjalanan berhasil diambil"
                    },
                    "404": {
                        "description": "Kontainer tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Gateway - Tracking Logs"
                ],
                "summary": "Tambah log perjalanan (Admin only)",
                "description": "Membuat log perjalanan baru untuk kontainer tertentu. Hanya admin.",
                "operationId": "createTrackingLog",
                "parameters": [
                    {
                        "name": "containerId",
                        "in": "path",
                        "description": "ID kontainer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "location_from",
                                    "location_to"
                                ],
                                "properties": {
                                    "location_from": {
                                        "type": "string",
                                        "example": "Gudang Utama - Jakarta Utara"
                                    },
                                    "location_to": {
                                        "type": "string",
                                        "example": "Zona B - Tangerang"
                                    },
                                    "status_change": {
                                        "type": "string",
                                        "example": "Active → Full"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Kontainer dipindahkan karena kapasitas penuh"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Log berhasil ditambahkan"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Kontainer tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan JWT token yang didapat dari endpoint login.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}